requirements.txt
# Core LLM libraries
google-generativeai

# For loading environment variables
python-dotenv

# For robust Windows UI Automation
pywinauto

# For colorized terminal output
colorama

# For web browser automation
selenium
webdriver-manager

# For sending WhatsApp messages
pywhatkit
